add ToDo - prep for live branch
To use GitHub for deployment requires a way to safely push from testing
to production without compromising the key or agent strings. By ignoring
the key file on the test and production copies, it won't ever get picked
up into github, and will already be in place on the production server.